Separating out HardwareMonitor from its listener.
It's usable separately.  Also this doesn't propagate the include
dependency on rocm_smi.h .
